# app/docs/openapi.yaml
openapi: 3.1.0
info:
  title: whisperX REST service
  description: "\n    # whisperX RESTful API\n\n    Welcome to the whisperX RESTful\
    \ API! This API provides a suite of audio processing services to enhance and analyze\
    \ your audio content.\n\n    ## Documentation:\n\n    For detailed information\
    \ on request and response formats, consult the [WhisperX Documentation](https://github.com/m-bain/whisperX).\n\
    \n    ## Services:\n\n    Speech-2-Text provides a suite of audio processing services\
    \ to enhance and analyze your audio content. The following services are available:\n\
    \n    1. Transcribe: Transcribe an audio/video  file into text.\n    2. Align:\
    \ Align the transcript to the audio/video file.\n    3. Diarize: Diarize an audio/video\
    \ file into speakers.\n    4. Combine Transcript and Diarization: Combine the\
    \ transcript and diarization results.\n\n    ## Supported file extensions:\n \
    \   AUDIO_EXTENSIONS = {'.aac', '.mp3', '.oga', '.m4a', '.wav', '.ogg', '.wma',\
    \ '.amr', '.awb', '.flac'}\n\n    VIDEO_EXTENSIONS = {'.avi', '.mkv', '.webm',\
    \ '.wmv', '.mp4', '.mov'}\n\n    ## Language-Specific Model Overrides:\n\n   \
    \ Some languages have fine-tuned models that are automatically used when that\
    \ language is selected.\n    The server-side `LANGUAGE_MODEL_OVERRIDES` configuration\
    \ maps language codes to optimized CTranslate2\n    model paths. When an override\
...
  version: 0.0.1
paths:
  /speech-to-text:
    post:
      tags:
      - Speech-2-Text
      summary: Speech To Text
      description: "Process an uploaded audio file for speech-to-text conversion.\n\
        \nSync ``def`` so FastAPI runs it in the threadpool: saving and hashing the\n\
        upload never block the event loop. Decoding runs in the background job.\n\n\
        Args:\n    background_tasks (BackgroundTasks): Background tasks dependency.\n\
        \    model_params (WhisperModelParams): Whisper model parameters.\n    align_params\
        \ (AlignmentParams): Alignment parameters.\n    diarize_params (DiarizationParams):\
        \ Diarization parameters.\n    asr_options_params (ASROptions): ASR options\
        \ parameters.\n    vad_options_params (VADOptions): VAD options parameters.\n\
        \    file (UploadFile): Uploaded audio file.\n    callback_url (str | None):\
        \ Optional URL to call back when processing is complete.\n    repository (ITaskRepository):\
        \ Task repository dependency.\n    file_service (FileService): File service\
        \ dependency.\n\nReturns:\n    Response: The queued task, or the in-flight\
        \ task an identical\n    earlier submit created."
      operationId: speech_to_text_speech_to_text_post
      parameters:
      - name: language
        in: query
        required: false
        schema:
          type: string
          description: Language to transcribe (ISO 639-1 code). Some languages have
            server-side fine-tuned model overrides (e.g. 'lv' uses a dedicated Latvian
            model with ~6x better accuracy). When an override is active, the 'model'
            parameter is ignored.
          enum:
          - en
          - zh
          - de
          - es
          - ru
          - ko
          - fr
          - ja
          - pt
          - tr
          - pl
          - ca
          - nl
          - ar
          - sv
          - it
          - id
          - hi
          - fi
          - vi
          - he
          - uk
          - el
          - ms
          - cs
          - ro
          - da
          - hu
          - ta
          - 'no'
          - th
          - ur
          - hr
          - bg
          - lt
          - la
          - mi
          - ml
          - cy
          - sk
          - te
          - fa
          - lv
          - bn
          - sr
          - az
          - sl
          - kn
          - et
          - mk
          - br
          - eu
          - is
          - hy
          - ne
          - mn
          - bs
          - kk
          - sq
          - sw
          - gl
          - mr
          - pa
          - si
          - km
          - sn
          - yo
          - so
          - af
          - oc
          - ka
          - be
          - tg
          - sd
          - gu
          - am
          - yi
          - lo
          - uz
          - fo
          - ht
          - ps
          - tk
          - nn
          - mt
          - sa
          - lb
          - my
          - bo
          - tl
          - mg
          - as
          - tt
          - haw
          - ln
          - ha
          - ba
          - jw
          - su
          - yue
          default: en
          title: Language
        description: Language to transcribe (ISO 639-1 code). Some languages have
          server-side fine-tuned model overrides (e.g. 'lv' uses a dedicated Latvian
          model with ~6x better accuracy). When an override is active, the 'model'
          parameter is ignored.
      - name: task
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/TaskEnum'
          description: Whether to perform X->X speech recognition ('transcribe') or
            X->English translation ('translate')
          default: transcribe
        description: Whether to perform X->X speech recognition ('transcribe') or
          X->English translation ('translate')
      - name: model
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/WhisperModel'
          description: Name of the Whisper model to use
          default: tiny
        description: Name of the Whisper model to use
      - name: device
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Device'
          description: Device to use for PyTorch inference
          default: cuda
        description: Device to use for PyTorch inference
      - name: device_index
        in: query
        required: false
        schema:
          type: integer
          description: Device index to use for FasterWhisper inference
          default: 0
          title: Device Index
        description: Device index to use for FasterWhisper inference
      - name: threads
        in: query
        required: false
        schema:
          type: integer
          description: Number of threads used by torch for CPU inference; supersedes
            MKL_NUM_THREADS/OMP_NUM_THREADS
          default: 0
          title: Threads
        description: Number of threads used by torch for CPU inference; supersedes
          MKL_NUM_THREADS/OMP_NUM_THREADS
      - name: batch_size
        in: query
        required: false
        schema:
          type: integer
          description: The preferred batch size for inference
          default: 8
          title: Batch Size
        description: The preferred batch size for inference
      - name: chunk_size
        in: query
        required: false
        schema:
          type: integer
          description: Chunk size for merging VAD segments. Default is 20, reduce
            this if the chunk is too long.
          default: 20
          title: Chunk Size
        description: Chunk size for merging VAD segments. Default is 20, reduce this
          if the chunk is too long.
      - name: compute_type
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/ComputeType'
          description: Type of computation
          default: float16
        description: Type of computation
      - name: align_model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Name of phoneme-level ASR model to do alignment
          title: Align Model
        description: Name of phoneme-level ASR model to do alignment
      - name: interpolate_method
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/InterpolateMethod'
          description: For word .srt, method to assign timestamps to non-aligned words,
            or merge them into neighboring.
          default: nearest
        description: For word .srt, method to assign timestamps to non-aligned words,
          or merge them into neighboring.
      - name: return_char_alignments
        in: query
        required: false
        schema:
          type: boolean
          description: Return character-level alignments in the output json file
          default: false
          title: Return Char Alignments
        description: Return character-level alignments in the output json file
      - name: min_speakers
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Minimum number of speakers to in audio file
          title: Min Speakers
        description: Minimum number of speakers to in audio file
      - name: max_speakers
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Maximum number of speakers to in audio file
          title: Max Speakers
        description: Maximum number of speakers to in audio file
      - name: beam_size
        in: query
        required: false
        schema:
          type: integer
          description: Number of beams in beam search, only applicable when temperature
            is zero
          default: 5
          title: Beam Size
        description: Number of beams in beam search, only applicable when temperature
          is zero
      - name: best_of
        in: query
        required: false
        schema:
          type: integer
          description: Number of beams to keep in beam search, only applicable when
            temperature is zero
          default: 5
          title: Best Of
        description: Number of beams to keep in beam search, only applicable when
          temperature is zero
      - name: patience
        in: query
        required: false
        schema:
          type: number
          description: Optional patience value to use in beam decoding
          default: 1.0
          title: Patience
        description: Optional patience value to use in beam decoding
      - name: length_penalty
        in: query
        required: false
        schema:
          type: number
          description: Optional token length penalty coefficient
          default: 1.0
          title: Length Penalty
        description: Optional token length penalty coefficient
      - name: temperatures
        in: query
        required: false
        schema:
          type: number
          description: Temperature to use for sampling
          default: 0.0
          title: Temperatures
        description: Temperature to use for sampling
      - name: compression_ratio_threshold
        in: query
        required: false
        schema:
          type: number
          description: If the gzip compression ratio is higher than this value, treat
            the decoding as failed
          default: 2.4
          title: Compression Ratio Threshold
        description: If the gzip compression ratio is higher than this value, treat
          the decoding as failed
      - name: log_prob_threshold
        in: query
        required: false
        schema:
          type: number
          description: If the average log probability is lower than this value, treat
            the decoding as failed
          default: -1.0
          title: Log Prob Threshold
        description: If the average log probability is lower than this value, treat
          the decoding as failed
      - name: no_speech_threshold
        in: query
        required: false
        schema:
          type: number
          description: If the probability of the token is higher than this value AND
            the decoding has failed due to `logprob_threshold`, consider the segment
            as silence
          default: 0.6
          title: No Speech Threshold
        description: If the probability of the token is higher than this value AND
          the decoding has failed due to `logprob_threshold`, consider the segment
          as silence
      - name: initial_prompt
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional text to provide as a prompt for the first window.
          title: Initial Prompt
        description: Optional text to provide as a prompt for the first window.
      - name: suppress_tokens
        in: query
        required: false
        schema:
          type: array
          items:
            type: integer
          description: Comma-separated list of token ids to suppress during sampling
          default:
          - -1
          title: Suppress Tokens
        description: Comma-separated list of token ids to suppress during sampling
      - name: suppress_numerals
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Whether to suppress numeric symbols and currency symbols during
            sampling
          default: false
          title: Suppress Numerals
        description: Whether to suppress numeric symbols and currency symbols during
          sampling
      - name: hotwords
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Hotwords related prompt applied before each transcription window
          title: Hotwords
        description: Hotwords related prompt applied before each transcription window
      - name: vad_onset
        in: query
        required: false
        schema:
          type: number
          description: Onset threshold for VAD (see pyannote.audio), reduce this if
            speech is not being detected
          default: 0.5
          title: Vad Onset
        description: Onset threshold for VAD (see pyannote.audio), reduce this if
          speech is not being detected
      - name: vad_offset
        in: query
        required: false
        schema:
          type: number
          description: Offset threshold for VAD (see pyannote.audio), reduce this
            if speech is not being detected.
          default: 0.363
          title: Vad Offset
        description: Offset threshold for VAD (see pyannote.audio), reduce this if
          speech is not being detected.
      - name: callback_url
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uri
            minLength: 1
            maxLength: 2083
          - type: 'null'
          title: Callback Url
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_speech_to_text_speech_to_text_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_speech_to_text_speech_to_text_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_speech_to_text_speech_to_text_post
...
    ComputeType:
      type: string
      enum:
      - float16
      - float32
      - int8
      title: ComputeType
      description: Enum for compute types.
...
    Device:
      type: string
      enum:
      - cuda
      - cpu
      title: Device
      description: Enum for device types.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InterpolateMethod:
      type: string
      enum:
      - nearest
      - linear
      - ignore
      title: InterpolateMethod
      description: Enum for interpolation methods.
...
    Response:
      properties:
        identifier:
          type: string
          title: Identifier
        message:
          type: string
          title: Message
      type: object
      required:
      - identifier
      - message
      title: Response
      description: Response model for API responses.
...
    TaskEnum:
      type: string
      enum:
      - transcribe
      - translate
      title: TaskEnum
      description: Enum for task types.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WhisperModel:
      type: string
      enum:
      - tiny
      - tiny.en
      - base
      - base.en
      - small
      - small.en
      - medium
      - medium.en
      - large
      - large-v1
      - large-v2
      - large-v3
      - large-v3-turbo
      - distil-large-v2
      - distil-medium.en
      - distil-small.en
      - distil-large-v3
      - nyrahealth/faster_CrisperWhisper
      title: WhisperModel
      description: Enum for Whisper model types.
...